server/requirements.txt
google-cloud-vision
google-cloud-translate
six
fastapi
starlette
pydantic
uvicorn